python-backend/requirements.txt
asgiref==3.3.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
DAWG-Python==0.7.2
defusedxml==0.7.0rc1
Django==3.1.3
django-cors-headers==3.5.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
djoser==2.1.0
docopt==0.6.2
drf-yasg==1.20.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
intervaltree==3.1.0
ipymarkup==0.9.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.17.0
loguru==0.5.3
MarkupSafe==1.1.1
natasha==1.4.0
navec==0.10.0
numpy==1.19.4
oauthlib==3.1.0
packaging==20.4
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==2.4.7
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2020.4
razdel==0.5.0
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
slovnet==0.5.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sortedcontainers==2.3.0
sqlparse==0.4.1
threadpoolctl==2.1.0
uritemplate==3.0.1
urllib3==1.25.11
win32-setctime==1.0.3
yargy==0.15.0
zipp==3.4.0
